Option to delay conflict checking of old pull requests until page view
[repository.pull-request] CHECK_ONLY_LAST_UPDATED_DAYS specifies the
number of days in which the PR must have been last updated for conflict
checking to happen immediately when the base branch is updated.

The default value of -1 behaves exactly as before with no delay. With 0
all conflict checking will be delayed until page view. We have found
that 1 day works well in practice.

This mostly eliminates the "Merge conflict checking is in progress. Try
again in few moments." message when conflict checking is slow for big
repositories and many pull requests.

PRs that are actively being worked on will be checked immediately, while
for others are likely to actually get checked in the next few seconds.

Even better would be to auto-update the page as GitHub does, but this
relatively simple change effectively solved the problem for us.